PUTA PRACTICA 3 FUNCIONAL

+++ P3/src/linked-list.h
@@ -39,7 +39,7 @@ typedef struct list_data_ {
   // within the structure. This additional information is associated
   // to the key. You may include any field you may need useful.
   int num_flights;
-  double delay;
+  int delay;
 } list_data;
